+++ examples/components/footer.vue
@@ -35,7 +35,7 @@
               class="footer__item">
         <p class="footer__title">更新日志</p>
         <p class="footer__product"></p>
-        <p class="footer__product">avue(v1.3.4) 更新日期：2018-11-15</p>
+        <p class="footer__product">avue(v1.4.0) 更新日期：2018-11-16</p>
         <p class="footer__product">avue-cli 更新日期：2018-11-11</p>
       </el-col>
     </el-row>