+++ packages/crud/src/crud-select.vue
@@ -16,6 +16,7 @@
                :value="item[valueKey]">
     </el-option>
   </el-select>
+
 </template>
 
 <script>